+++ CHANGELOG.md
@@ -12,6 +12,10 @@ N/A
 
 #### Enhancements
 
+- Added Data extensions
+    - data to hex string
+    - hex string to data
+    - data to UInt8 array
 - Added SystemUtility struct with shell command access
 - added ip address validation extension to String